refactor PreviousTransitions
